Remove inline `CombineCSVs` class
This is now replaced by the new `DataProcessor` class.
