+++ lib/data_processor.rb
@@ -117,4 +117,53 @@ class DataProcessor
     sunday = (Date.parse(monday) + 6).strftime("%d/%m/%Y")
     [saturday, sunday]
   end
+
+  def self.validate_responses(responses_csv:)
+    responses_data = CSV.read(responses_csv, headers: true)
+    columns = responses_data.headers
+    first_columns_regexes = [
+      /^Timestamp$/,
+      /^Email address$/,
+      /^Have you been given an exemption from on call\?/,
+      /^Do you have any non working days\?/,
+      /^What team\/area are you in/,
+      /^If you work different hours to the 9.30am-5.30pm 2nd line shifts, please state your hours/,
+    ]
+    week_commencing_regex = /^Week commencing \d+{2}\/\d{2}\/\d{4}$/
+    last_column_regex = /^Need to elaborate on any of the above\?/
+
+    first_columns = columns.shift(first_columns_regexes.count).each_with_index.map do |column, index|
+      { regex: first_columns_regexes[index], value: column }
+    end
+    last_column = [columns.pop].map { |column| { regex: last_column_regex, value: column } }
+    week_columns = columns.map { |column| { regex: week_commencing_regex, value: column } }
+
+    (first_columns + week_columns + last_column).each do |hash|
+      unless hash[:value].match(hash[:regex])
+        raise InvalidStructureException, "Expected '#{hash[:value]}' to match '#{hash[:regex]}'"
+      end
+    end
+
+    validate_week_dates(week_columns.map { |hash| hash[:value].sub("Week commencing ", "") })
+
+    true
+  end
+
+  def self.validate_week_dates(week_dates)
+    date_this_week = week_dates.shift(1).first
+    day_of_week = Date.parse(date_this_week).strftime("%A")
+
+    unless day_of_week == "Monday"
+      raise InvalidStructureException, "Expected column 'Week commencing #{date_this_week}' to correspond to a Monday, but it's a #{day_of_week}."
+    end
+
+    week_dates.each do |date_next_week|
+      expected_next_date = (Date.parse(date_this_week) + 7).strftime("%d/%m/%Y")
+      unless date_next_week == expected_next_date
+        raise InvalidStructureException, "Expected 'Week commencing #{date_next_week}' to be 'Week commencing #{expected_next_date}'"
+      end
+
+      date_this_week = date_next_week
+    end
+  end
 end
